fix : [피드백 반영] UserRepository에 포함되었던 기능 user/wishlist/cart/auth/order로 분리
UserRepository에는 유저 정보관련 기능만 포함하도록 수정

issue [#4]
