+++ app/src/main/java/com/example/commerceapp/domain/repository/CartRepository.kt
@@ -1,0 +1,4 @@
+package com.example.commerceapp.domain.repository
+
+interface CartRepository {
+}
